requirements.txt
# Core Language Models and RAG Components
langchain>=0.1.0
langchain-community>=0.0.9
langchain-openai>=0.0.3
langchain-ollama>=0.0.1

# Vector Store and Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.15.1
python-docx>=0.8.11
docx2txt>=0.8


# Vietnamese NLP
underthesea>=6.2.0
unidecode>=1.3.6
transformers>=4.33.1

# API and Server
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0

# Utilities
tqdm>=4.66.1
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0

# Logging and Monitoring
loguru>=0.7.0

# Additional dependencies
gradio>=4.0.0
redis>=5.0.0


